separate article list and article components, clean up blog component
